Deleteconsent generic methods relocated to OBModuleHelper
