+++ amorphie.consent/Helper/OBModuleHelper.cs
@@ -7,6 +7,8 @@ using amorphie.consent.core.DTO.OpenBanking.HHS;
 using amorphie.consent.core.DTO.Tag;
 using amorphie.consent.core.DTO.Token;
 using amorphie.consent.core.Enum;
+using amorphie.consent.core.Model;
+using amorphie.consent.data;
 using amorphie.consent.Service.Interface;
 using amorphie.consent.Service.Refit;
 using Jose;
@@ -371,7 +373,155 @@ public static class OBModuleHelper
         // Join the links with commas and set the header
         httpContext.Response.Headers["Link"] = string.Join(", ", links);
     }
+    
+   
+    /// <summary>
+    /// Cancels institution ohk type consent
+    /// </summary>
+    /// <returns>Cancel process result</returns>
+    public static async Task<ApiResult> CancelInstitutionConsentUnAuthorized(ConsentDbContext context, ITokenService tokenService,
+        IOBEventService eventService, IYosInfoService yosInfoService, Consent entity, string cancelDetailCode)
 
+    {
+        ApiResult result = new();
+        //State list can be cancelled from login
+        var canBeCancelledStates = new List<string>()
+        {
+            OpenBankingConstants.RizaDurumu.YetkiBekleniyor,
+            OpenBankingConstants.RizaDurumu.Yetkilendirildi,
+            OpenBankingConstants.RizaDurumu.YetkiKullanildi
+        };
+       
+        if (!canBeCancelledStates.Contains(entity.State))
+        {
+            result.Result = false;
+            result.Message = "Consent state not valid to be cancelled.";
+            return result;
+        }
+
+        //TODO:EKip iptal detay kodu uygun bir kod ile set edilmeli. OpenBankingConstants.RizaIptalDetayKodu.GKDIptali_OHKHHSKontrolleriniAsamadi
+        return await CancelConsent(context, tokenService, eventService, yosInfoService, entity,cancelDetailCode);
+    }
+    
+    /// <summary>
+    /// Cancel consent generic method to cancel consent without any control.
+    /// Checks consent type and cancels according to consent type
+    /// </summary>
+    /// <returns></returns>
+    public static async Task<ApiResult> CancelConsent(ConsentDbContext context, ITokenService tokenService,
+        IOBEventService eventService, IYosInfoService yosInfoService, Consent entity, string cancelDetailCode )
+    {
+        ApiResult result = new();
+        if (entity.ConsentType == ConsentConstants.ConsentType.OpenBankingAccount)
+        {
+            //Account consent
+            //Update consent rıza bilgileri properties
+            await  CancelAccountConsent(context, tokenService, eventService: eventService, yosInfoService, entity,
+                cancelDetailCode);
+        }
+        else if (entity.ConsentType == ConsentConstants.ConsentType.OpenBankingPayment)
+        {
+            //Payment consent
+            //Update consent rıza bilgileri properties
+            await  CancelPaymentConsent(context, tokenService, entity,
+                cancelDetailCode);
+        }
+        else
+        {
+            //Not related type
+            result.Result = false;
+            result.Message = "Consent type not valid";
+        }
+        return result;
+    }
+
+
+
+    /// <summary>
+    /// Cancels account consent.
+    /// Updates additionaldata, consent state, send to account serive clms.
+    /// If there is token, revokes the token
+    /// If event subscription, send event to yos
+    /// </summary>
+    public static async Task CancelAccountConsent(ConsentDbContext context, ITokenService tokenService,
+        IOBEventService eventService, IYosInfoService yosInfoService, Consent entity, string cancelDetailCode )
+    {
+       var currentState = entity.State;//current consent state
+        //Update consent rıza bilgileri properties
+        var additionalData = JsonSerializer.Deserialize<HesapBilgisiRizasiHHSDto>(entity!.AdditionalData);
+        additionalData!.rzBlg.rizaDrm = OpenBankingConstants.RizaDurumu.YetkiIptal;
+        additionalData.rzBlg.rizaIptDtyKod = cancelDetailCode;
+        additionalData.rzBlg.gnclZmn = DateTime.UtcNow;
+        entity.AdditionalData = JsonSerializer.Serialize(additionalData);
+        entity.ModifiedAt = DateTime.UtcNow;
+        entity.State = OpenBankingConstants.RizaDurumu.YetkiIptal;
+        entity.StateModifiedAt = DateTime.UtcNow;
+        entity.StateCancelDetailCode = additionalData.rzBlg.rizaIptDtyKod;
+        //Update consent detail to send consent information to account service.
+        var consentDetail = entity.OBAccountConsentDetails.FirstOrDefault();
+        if (consentDetail is not null)
+        {
+            consentDetail.SendToServiceTryCount = 0;
+            consentDetail.SendToServiceDeliveryStatus = OpenBankingConstants.RecordDeliveryStatus.Processing;
+            context.OBAccountConsentDetails.Update(consentDetail);
+        }
+
+        context.Consents.Update(entity);
+        await context.SaveChangesAsync();
+
+        if (currentState == OpenBankingConstants.RizaDurumu.YetkiKullanildi)
+        {
+            //Revoke token
+            await tokenService.RevokeConsentToken(entity.Id);
+        }
+        
+        //If YOS has subscription, Do event process
+        ApiResult yosHasSubscription = await yosInfoService.IsYosSubscsribed(entity.Variant!,
+            OpenBankingConstants.OlayTip.KaynakGuncellendi, OpenBankingConstants.KaynakTip.HesapBilgisiRizasi);
+        if (yosHasSubscription.Result
+            && yosHasSubscription.Data != null
+            && (bool)yosHasSubscription.Data)
+        {
+
+            //Send event to yos
+            await eventService.DoEventProcess(entity.Id.ToString(),
+                additionalData.katilimciBlg,
+                eventType: OpenBankingConstants.OlayTip.KaynakGuncellendi,
+                sourceType: OpenBankingConstants.KaynakTip.HesapBilgisiRizasi,
+                sourceNumber: entity.Id.ToString());
+        }
+    }
+
+    
+    /// <summary>
+    /// Cancels payment consent.
+    /// Updates additionaldata, consent state
+    /// </summary>
+      public static async Task CancelPaymentConsent(ConsentDbContext context, ITokenService tokenService,
+          Consent entity, string cancelDetailCode )
+    {
+       var currentState = entity.State;//current consent state
+        //Update consent rıza bilgileri properties
+        var additionalData = JsonSerializer.Deserialize<OdemeEmriRizasiWithMsrfTtrHHSDto>(entity.AdditionalData);
+        additionalData!.rzBlg.rizaDrm = OpenBankingConstants.RizaDurumu.YetkiIptal;
+        additionalData.rzBlg.rizaIptDtyKod = cancelDetailCode;
+        additionalData.rzBlg.gnclZmn = DateTime.UtcNow;
+        entity.AdditionalData = JsonSerializer.Serialize(additionalData);
+        entity.ModifiedAt = DateTime.UtcNow;
+        entity.State = OpenBankingConstants.RizaDurumu.YetkiIptal;
+        entity.StateModifiedAt = DateTime.UtcNow;
+        entity.StateCancelDetailCode = additionalData.rzBlg.rizaIptDtyKod;
+        context.Consents.Update(entity);
+        await context.SaveChangesAsync();
+
+        if (currentState == OpenBankingConstants.RizaDurumu.YetkiKullanildi)
+        {
+            //Revoke token
+            await tokenService.RevokeConsentToken(entity.Id);
+        }
+    }
+
+  
 
 
 }
